KinectSecurity Project Ready!
The Kinect Security application is "ready". The solution is in the
folder KinectSecurity. You can now run the project by pressing CTRL+F5,
which will bring up the familiar demo that we have been playing with. I
left that demo in for testing purposes, but in the future we will need
to remove it, or replace it with something more appropriate.